Refactoring: Gradle, change rogig(parameter for remove another dependencies)
