+++ build.gradle
@@ -28,22 +28,25 @@ compileJava {
 			'-Amapstruct.suppressGeneratorTimestamp=true',   // 생성 코드에 타임스탬프 추가 방지
 			'-Amapstruct.suppressGeneratorVersionInfoComment=true',  // 생성 코드에 버전 정보 주석 추가 방지
 			'-Amapstruct.verbose=true',   // MapStruct의 주요 동작 로그 출력
-			"-Amapstruct.defaultComponentModel=spring" // Spring 환경이라면 추가
+			'-Amapstruct.defaultComponentModel=spring',  // MapStruct의 기본 컴포넌트 모델을 Spring으로 설정
 	]
 }
 
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation "org.mapstruct:mapstruct:1.6.3"
+	implementation 'org.mapstruct:mapstruct:1.5.5.Final'
+	implementation 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
 
 	compileOnly 'org.projectlombok:lombok'
 
 	annotationProcessor 'org.projectlombok:lombok'
-	annotationProcessor "org.mapstruct:mapstruct-processor:1.6.3"
+	annotationProcessor 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
+	annotationProcessor 'org.mapstruct:mapstruct-processor:1.5.5.Final'
+
 	runtimeOnly 'com.h2database:h2'
 	runtimeOnly 'org.postgresql:postgresql'
-	testAnnotationProcessor "org.mapstruct:mapstruct-processor:1.6.3"
+	testAnnotationProcessor "org.mapstruct:mapstruct-processor:1.5.5.Final"
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
