+++ src/main/java/team7/hrbank/domain/binary/LocalBinaryContentStorage.java
@@ -23,7 +23,6 @@ import static org.springframework.http.HttpHeaders.*;
 @Repository
 public class LocalBinaryContentStorage {
 
-    // 예외 자세한 처리는 나중에 ㄱ
     private final Path root;
 
     public LocalBinaryContentStorage(@Value("${hrbank.storage.local.root-path}") String root) {
@@ -31,9 +30,7 @@ public class LocalBinaryContentStorage {
         init();// 임시로
     }
 
-    //    public void put(byte[] content, BinaryContent binaryContent) {
     public void put(byte[] content, Long id) {
-        // 부서별 파일 이름 필요?? (나중에 의논 Cuz 부서별로 디렉토리 만들어서 저장? 부서에다 직원 ID 이렇게 커스텀하면 나중에 download 성능이 느려지거나 여러 파라미터 정보가 필요할 것)
         try {
             Files.write(resolvePath(id), content);
         } catch (Exception e) {
@@ -41,7 +38,7 @@ public class LocalBinaryContentStorage {
         }
     }
 
-    public ResponseEntity<Resource> download(Long id, Employee employee) {
+    public ResponseEntity<Resource> download(Long id) {
         Path profilePath = resolvePath(id);
         String fileType = MediaType.APPLICATION_OCTET_STREAM_VALUE;
         try {
@@ -54,7 +51,7 @@ public class LocalBinaryContentStorage {
         }
 
         HttpHeaders headers = new HttpHeaders();
-        headers.add(CONTENT_DISPOSITION, "attachment; filename=[" + employee.getDepartment().getName() + "]" + "_" + employee.getId() + "." + fileType.split("/")[0]);
+        headers.add(CONTENT_DISPOSITION, "attachment; filename= "+ profilePath.getFileName());
         headers.add(CONTENT_TYPE, fileType);
 
         return ResponseEntity
@@ -69,19 +66,8 @@ public class LocalBinaryContentStorage {
     private Path resolvePath(Long id) {
         return root.resolve(id.toString());
     }
-    //    private Path resolvePathForDownLoad(Employee employee) {
-//        //파일 저장 위치 규칙 예시: {root}/[부서명]_[사원ID].jpg
-//        // 이렇게 하면 불편한 점 : 부서정보, 타입정보, 사원정보, binaryContent정보가 필요함
-    // 파라미터로 받아올게 많아져서 불편 예상
-    // startwith 으로 필터링할 경우 전부 읽어오기 때문에 부담 예상
-//        String departmentName = employee.getDepartment().getName();
-//        Long id = employee.getId();
-//        return root.resolve(binaryContent.getId().toString() + "." + binaryContent.getFileType());
-//        return root.resolve("[" +departmentName + "]_" + id + ".jpg");
-//    }
 
     //루트 디렉토리를 초기화합니다.
-    //Bean이 생성되면 자동으로 호출되도록
     void init() {
         if (!Files.exists(root)) {
             try {
