+++ src/main/java/team7/hrbank/domain/binary/dto/BinaryContentDto.java
@@ -1,4 +1,4 @@
-package team7.hrbank.domain.binary;
+package team7.hrbank.domain.binary.dto;
 
 public record BinaryContentDto(
         String fileName,
